Move Ui interface to CommandLine folder
